+++ gluuapi/__init__.py
@@ -3,4 +3,4 @@
 #
 # All rights reserved.
 
-__version__ = "0.4.2-dev"
+__version__ = "0.4.3-dev"