+++ tests/setup/test_oxtrust_setup.py
@@ -1,5 +1,5 @@
 def test_setup(oxtrust_setup, patched_salt, patched_sleep,
-               salt_event_ok):
+               salt_event_ok, patched_run):
     # TODO: it might be better to split the tests
     oxtrust_setup.setup()
 
@@ -8,5 +8,5 @@ def test_delete_web_cert(oxtrust_setup, patched_salt):
     oxtrust_setup.delete_nginx_cert()
 
 
-def test_teardown(oxtrust_setup, patched_salt):
+def test_teardown(oxtrust_setup, patched_salt, patched_run):
     oxtrust_setup.teardown()